// src/models/World.ts
export interface World {
    updated: number
    cases: number
    todayCases: number
    deaths: number
    todayDeaths: number
    recovered: number
    active: number
    critical: number
    casesPerOneMillion: number
    deathsPerOneMillion: number
    tests: number
    testsPerOneMillion: number
    population: number
    activePerOneMillion: number
    recoveredPerOneMillion: number
    criticalPerOneMillion: number
    affectedCountries: number
}